% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemoDiv.R
\name{chemoDiv}
\alias{chemoDiv}
\title{FUNCTION_TITLE}
\usage{
chemoDiv(chem_dataset)
}
\arguments{
\item{chem_dataset}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
