% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{DOM molecules data}
\format{
The format for DOM data is a matrix with samples in the rows and DOM molecules in the columns.
}
\usage{
data
}
\description{
Example data of a DOM compositional table (56 samples by 7116 DOM known molecules)
}
\examples{
data(data)
}
\references{
Hu A, Choi M, Tanentzap AJ, Liu J, Jang K-S, Lennon JT, Liu Y, Soininen J, Lu X, Zhang Y, Shen J, Wang J. 2021.
Quantifying microbial associations of dissolved organic matter under global change. \emph{bioRxiv}.

Hu A, Choi M, Tanentzap AJ, Liu J, Jang K-S, Lennon JT, Liu Y, Soininen J, Lu X, Zhang Y, Shen J, Wang J. 2022.
Quantifying the associations between dissolved organic matter and microbes under global change. \emph{Nature communications}.
}
\keyword{datasets}
