% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{envi56}
\alias{envi56}
\title{Environmental data}
\format{
Environmental dataframe for 56 samples
}
\usage{
envi56
}
\description{
Environmental data
}
\examples{
data(envi56)
}
\keyword{datasets}
