% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_moltrait}
\alias{Example_moltrait}
\title{DOM traits data}
\format{
Trait information for DOM in Example_molcomm
}
\usage{
Example_moltrait
}
\description{
DOM traits data
}
\examples{
data(Example_moltrait)
}
\keyword{datasets}
