% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDME.R
\name{iDME}
\alias{iDME}
\title{FUNCTION_TITLE}
\usage{
iDME(
  comm_dom,
  know_seq,
  n = length(know_seq),
  bootstrap = 100,
  threshold = 0.3
)
}
\arguments{
\item{comm_dom}{PARAM_DESCRIPTION}

\item{know_seq}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION, Default: length(know_seq)}

\item{bootstrap}{PARAM_DESCRIPTION, Default: 100}

\item{threshold}{PARAM_DESCRIPTION, Default: 0.3}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[SpiecEasi]{sparcc}}
\code{\link[igraph]{graph_from_adjacency_matrix}}, \code{\link[igraph]{degree}}
}
