% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mol.all}
\alias{mol.all}
\title{DOM traits data}
\format{
Trait dataframe for 14853 DOM molecules(7116 known and 7737dark)
}
\usage{
mol.all
}
\description{
DOM traits data
}
\examples{
data(mol.all)
}
\keyword{datasets}
