% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molDK.R
\name{molDK}
\alias{molDK}
\title{FUNCTION_TITLE}
\usage{
molDK(dom_comm, dom_trait, metric = c("MolForm"))
}
\arguments{
\item{dom_comm}{PARAM_DESCRIPTION}

\item{dom_trait}{PARAM_DESCRIPTION}

\item{metric}{PARAM_DESCRIPTION, Default: c("MolForm")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{case_when}}
}
