% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIGLAS-package.R
\docType{package}
\name{NIGLAS-package}
\alias{NIGLAS}
\alias{NIGLAS-package}
\title{NIGLAS: What the Package Does (One Line, Title Case)}
\description{
Distinguish dark matter from microbial or dissolved organic matter community.What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
