% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDis.R
\name{sampleDis}
\alias{sampleDis}
\title{FUNCTION_TITLE}
\usage{
sampleDis(dom_comm, method = "euclidean")
}
\arguments{
\item{dom_comm}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION, Default: 'euclidean'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[vegan]{vegdist}}
}
