% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemoDiv.R
\name{chemoDiv}
\alias{chemoDiv}
\title{FUNCTION_TITLE}
\usage{
chemoDiv(chem_dataset)
}
\arguments{
\item{chem_dataset}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
#EXAMPLE1
# Example data of a DOM compositional table (50 samples by 200 DOM molecules)
data(data)
# Calculate selected types of α-diversity and evenness measures
chemoDiv(data)
}
}
