% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDME.R
\name{iDME}
\alias{iDME}
\title{FUNCTION_TITLE}
\usage{
iDME(
  comm_dom,
  know_seq,
  n = length(know_seq),
  bootstrap = 100,
  occ.t = 0.3,
  cor.t = 0.3
)
}
\arguments{
\item{comm_dom}{PARAM_DESCRIPTION}

\item{know_seq}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION, Default: length(know_seq)}

\item{bootstrap}{PARAM_DESCRIPTION, Default: 100}

\item{threshold}{PARAM_DESCRIPTION, Default: 0.3}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
 #EXAMPLE1
 # Example data of a DOM compositional table (50 samples by 200 DOM molecules)
 data(Example_domcomm)
 # Example data of a DOM trait table
 data(Example_domtrait)
 # Select DOMs with well-defined chemical compositions
 know_seq <- Example_domtrait$MolForm[!is.na(Example_domtrait$El_comp)]
 # Calculate iDEM value
 idme <- iDME(comm_dom = Example_domcomm,
             know_seq = know_seq,
             n = 50,
             bootstrap = 10,
             threshold = 0.03)
}
}
\seealso{
\code{\link[SpiecEasi]{sparcc}}
\code{\link[igraph]{graph_from_adjacency_matrix}}, \code{\link[igraph]{degree}}
}
